This is the database implementation and completion

+++ src/main/java/com/satish/shopsy/productservice8aug/Service/ProductService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface ProductService {
 
-    public Product getProductById(Long id);
+    public Product getProductById(Integer id);
 
 
     public Product createProduct(String title,
